Add separate controller files to all resources
