+++ vidly/routes/genres.js
@@ -1,47 +1,24 @@
-const express = require('express');
+const { Router } = require('express');
 const auth = require('../middleware/auth');
 const admin = require('../middleware/admin');
+const {
+  show,
+  index,
+  create,
+  update,
+  destroy,
+} = require('../controllers/genres');
 
-const router = express.Router();
+const router = Router();
 
-const { Genre, validate } = require('../models/genre');
+router.get('/', index);
 
-router.get('/', async (req, res) => {
-  const genres = await Genre.find().sort('name');
-  return res.send(genres);
-});
+router.post('/', auth, create);
 
-router.post('/', auth, async (req, res) => {
-  const { error } = validate(req.body);
-  if (error) return res.status(400).send(error.details[0].message);
-  const genre = new Genre({ name: req.body.name });
-  await genre.save();
-  return res.send(genre);
-});
+router.put('/:id', update);
 
-router.put('/:id', async (req, res) => {
-  const { error } = validate(req.body);
-  if (error) return res.status(400).send(error.details[0].message);
-  const genre = await Genre.updateOne(
-    { _id: req.params.id },
-    { name: req.body.name },
-    { new: true },
-  );
+router.delete('/:id', [auth, admin], destroy);
 
-  if (!genre) return res.status(404).send('The genre with the given ID was not found.');
-  return res.send(genre);
-});
-
-router.delete('/:id', [auth, admin], async (req, res) => {
-  const genre = await Genre.deleteOne({ _id: req.params.id });
-  if (!genre) return res.status(404).send('The genre with the given ID was not found.');
-  return res.send(genre);
-});
-
-router.get('/:id', async (req, res) => {
-  const genre = await Genre.findById(req.params.id);
-  if (!genre) return res.status(404).send('The genre with the given ID was not found.');
-  return res.send(genre);
-});
+router.get('/:id', show);
 
 module.exports = router;
