+++ vidly/routes/movies.js
@@ -1,72 +1,23 @@
-const express = require('express');
-const { Movie, validate } = require('../models/movie');
-const { Genre } = require('../models/genre');
+const { Router } = require('express');
 
-const router = express.Router();
+const {
+  show,
+  index,
+  create,
+  update,
+  destroy,
+} = require('../controllers/movies');
 
-router.get('/', async (req, res) => {
-  const movies = await Movie.find().sort('name');
-  return res.send(movies);
-});
+const router = Router();
 
-router.post('/', async (req, res) => {
-  const { error } = validate(req.body);
-  if (error) return res.status(400).send(error.details[0].message);
+router.get('/', index);
 
-  const genre = await Genre.findById(req.body.genreId);
-  if (!genre) return res.status(400).send('Invalid genre.');
+router.post('/', create);
 
-  const movie = new Movie({
-    title: req.body.title,
-    genre: {
-      _id: genre._id,
-      name: genre.name,
-    },
-    numberInStock: req.body.numberInStock,
-    dailyRentalRate: req.body.dailyRentalRate,
-  });
-  await movie.save();
+router.put('/:id', update);
 
-  return res.send(movie);
-});
+router.delete('/:id', destroy);
 
-router.put('/:id', async (req, res) => {
-  const { error } = validate(req.body);
-  if (error) return res.status(400).send(error.details[0].message);
-
-  const genre = await Genre.findById(req.body.genreId);
-  if (!genre) return res.status(400).send('Invalid genre.');
-
-  const movie = await Movie.updateOne({ _id: req.params.id },
-    {
-      title: req.body.title,
-      genre: {
-        _id: genre._id,
-        name: genre.name,
-      },
-      numberInStock: req.body.numberInStock,
-      dailyRentalRate: req.body.dailyRentalRate,
-    });
-
-  if (!movie) return res.status(404).send('The movie with the given ID was not found.');
-
-  return res.send(movie);
-});
-
-router.delete('/:id', async (req, res) => {
-  const movie = await Movie.deleteOne({ _id: req.params.id });
-
-  if (!movie) return res.status(404).send('The movie with the given ID was not found.');
-
-  return res.send(movie);
-});
-
-router.get('/:id', async (req, res) => {
-  const movie = await Movie.findById(req.params.id);
-
-  if (!movie) return res.status(404).send('The movie with the given ID was not found.');
-
-  return res.send(movie);
-});
+router.get('/:id', show);
 
 module.exports = router;
