+++ vidly/routes/rentals.js
@@ -1,63 +1,16 @@
-const mongoose = require('mongoose');
-const Fawn = require('fawn');
-const express = require('express');
-const { Rental, validate } = require('../models/rental');
-const { Movie } = require('../models/movie');
-const { Customer } = require('../models/customer');
+const { Router } = require('express');
+const {
+  show,
+  index,
+  create,
+} = require('../controllers/rentals');
 
-const router = express.Router();
+const router = Router();
 
-Fawn.init(mongoose);
+router.get('/', index);
 
-router.get('/', async (req, res) => {
-  const rentals = await Rental.find().sort('-dateOut');
-  return res.send(rentals);
-});
+router.post('/', create);
 
-router.post('/', async (req, res) => {
-  const { error } = validate(req.body);
-  if (error) return res.status(400).send(error.details[0].message);
-
-  const customer = await Customer.findById(req.body.customerId);
-  if (!customer) return res.status(400).send('Invalid customer.');
-
-  const movie = await Movie.findById(req.body.movieId);
-  if (!movie) return res.status(400).send('Invalid movie.');
-
-  if (movie.numberInStock === 0) return res.status(400).send('Movie out of stock!');
-
-  const rental = new Rental({
-    customer: {
-      _id: customer._id,
-      name: customer.name,
-      phone: customer.phone,
-    },
-    movie: {
-      _id: movie._id,
-      title: movie.title,
-      dailyRentalRate: movie.dailyRentalRate,
-    },
-  });
-
-  try {
-    const task = Fawn.Task();
-    task.save('rentals', rental)
-      .update('movies', { _id: movie._id }, {
-        $inc: { numberInStock: -1 },
-      })
-      .run();
-    return res.send(rental);
-  } catch (error) {
-    return res.status(500).send('Something went wrong');
-  }
-});
-
-router.get('/:id', async (req, res) => {
-  const rental = await Rental.findById(req.params.id);
-
-  if (!rental) return res.status(404).send('The rental with the given ID was not found.');
-
-  return res.send(rental);
-});
+router.get('/:id', show);
 
 module.exports = router;
